// src/app/interfaces/interfaces.ts
export interface dPuntuaciones {
    parada: number
    paso: number
    trote: number
    galope: number
    pasoAtras: number
    transiciones: number
    cambioDirec: number
    figuras: number
    movLateral: number
    piruetas: number
}

export interface sPenalizaciones {
    derriboS: number
    rehusoS: number
    caidaS: number
    tiempoS: number
}

export interface cPenalizaciones {
    rehusoC: number
    caidaC: number
    tiempoC: number
}